requirements.txt
flask==2.0.1
Pillow==8.3.2
werkzeug==2.0.1
pytest==6.2.5
tensorflow==2.9.1
keras==2.9.0
numpy==1.23.1
beautifulsoup4==4.11.1
requests==2.28.1
scikit-learn==1.1.2
nltk==3.7
selenium==4.3.0
webdriver_manager==3.8.0